[BE] Fix: 회원탈퇴 오류 수정

See merge request s11-final/S11P31D105!177

+++ backend/src/main/java/com/checkitout/ijoa/child/service/ChildrenManagementService.java
@@ -197,6 +197,10 @@ public class ChildrenManagementService {
 
         String profileUrl = child.getProfile();
 
+        if (profileUrl == null || profileUrl.isEmpty()) {
+            return;
+        }
+
         if (!profileUrl.equals(GIRL_PROFILE_DEFAULT_URL) && !profileUrl.equals(BOY_PROFILE_DEFAULT_URL)) {
             fileService.deleteFile(getKeyFromUrl(child.getProfile()));
         }
